webhook/requirements.txt
Flask
gunicorn
google-cloud-pubsub
google-cloud-secret-manager